README for Movie recommendation
The Movie Recommendation System is a Python-based project that suggests movies to users based on similarity in movie genres. It uses cosine similarity and vectorization to calculate how closely one movie resembles another. By taking a movie title as input, the system analyzes genre patterns and returns the top most similar movies, helping users easily discover films they are likely to enjoy. The purpose of this project is to understand the basics of recommendation systems and how machine learning can be applied to content-based filtering.